Create routes for inference
