+++ digits/inference/images/classification/__init__.py
@@ -1,5 +1,4 @@
 # Copyright (c) 2016, NVIDIA CORPORATION.  All rights reserved.
 from __future__ import absolute_import
 
-from .classification import *
-from .job import ImageInferenceJob
+from .job import ImageInferenceClassifyManyJob
