+++ digits/webapp.py
@@ -51,6 +51,10 @@ import digits.model.images.classification.views
 app.register_blueprint(digits.model.images.classification.views.blueprint, url_prefix='/models/images/classification')
 import digits.model.images.generic.views
 app.register_blueprint(digits.model.images.generic.views.blueprint, url_prefix='/models/images/generic')
+import digits.inference.views
+app.register_blueprint(digits.inference.views.blueprint, url_prefix='/inference')
+import digits.inference.images.classification.views
+app.register_blueprint(digits.inference.images.classification.views.blueprint, url_prefix='/inference/images/classification')
 
 def username_decorator(f):
     from functools import wraps
